+++ packages/impulse/lib/src/error_handling.dart
@@ -1,30 +1,35 @@
-import 'package:signals_flutter/signals_flutter.dart';
-
-typedef Result<T> = (T?, AsyncError<T>?);
+typedef Result<T> = (T? value, Err? err);
 
 Future<Result<T>> attempt<T>(Future<T> Function() call) async {
   try {
     return (await call(), null);
   } catch (e, st) {
-    return (null, AsyncError<T>(e, st));
+    return (null, Err(e, st));
   }
+}
+
+class Err {
+  Err(this.error, this.stackTrace);
+
+  final Object error;
+  final StackTrace stackTrace;
 }
 
 extension MapResult<T> on Result<T> {
   R map<R>({
     required R Function() onNothing,
     required R Function(T value) onValue,
-    required R Function(AsyncError err) onError,
-    R Function(T value, AsyncError err)? onValueAndError,
+    required R Function(Err err) onError,
+    R Function(T value, Err err)? onValueAndError,
   }) {
     return switch (this) {
       (null, null) => onNothing(),
-      (T value, AsyncError err) when onValueAndError != null => onValueAndError(
+      (T value, Err err) when onValueAndError != null => onValueAndError(
         value,
         err,
       ),
       (T value, _) => onValue(value),
-      (_, AsyncError err) => onError(err),
+      (_, Err err) => onError(err),
     };
   }
 }