+++ packages/impulse/lib/src/reference.dart
@@ -17,6 +17,11 @@ typedef FamilyCreate<T, R> = T Function(Store store, R param);
 /// A function that retrieves an instance of [T].
 typedef Get<T> = T Function();
 
+/// Interface for references that dont have a key (so all of them except familyRef)
+abstract class KeylessRef<T> {
+  ImpulseReference<T> call();
+}
+
 /// A definition of how an object is created, identified, and managed within the store.
 @immutable
 class ImpulseReference<T> {
@@ -53,7 +58,7 @@ class ImpulseReference<T> {
 
 /// A reference that creates a new instance every time it is retrieved from the store.
 @immutable
-class FactoryRef<T> {
+class FactoryRef<T> implements KeylessRef<T> {
   /// Creates a [FactoryRef] with a [create] function and an optional [dispose] function.
   const FactoryRef(this.create, {this.reassemble, this.dispose});
 
@@ -76,12 +81,13 @@ class FactoryRef<T> {
   }
 
   /// Returns an [ImpulseReference] representing this factory.
+  @override
   ImpulseReference<T> call() {
     return ImpulseReference(
       key: this,
       create: create,
       isFactory: true,
-      keepAlive: false,
+      keepAlive: true,
       reassemble: reassemble,
       dispose: dispose,
     );
@@ -90,15 +96,9 @@ class FactoryRef<T> {
 
 /// A reference that creates and caches a single instance in the store.
 @immutable
-class Ref<T> {
+class Ref<T> implements KeylessRef<T> {
   /// Creates a [Ref] with a [create] function.
-  /// If [keepAlive] is true, the object will not be automatically dropped when not in use.
-  const Ref(
-    this.create, {
-    this.reassemble,
-    this.dispose,
-    this.keepAlive = false,
-  });
+  const Ref(this.create, {this.reassemble, this.dispose});
 
   /// The function used to create the object.
   final Create<T> create;
@@ -111,8 +111,44 @@ class Ref<T> {
   /// An optional function to handle manual disposal of the object.
   final Dispose<T>? dispose;
 
-  /// Whether the object should remain in the store even when its reference count reaches zero.
-  final bool keepAlive;
+  /// Get the value this [Ref] refers to from the [store].
+  T get(Store store) {
+    final ref = call();
+
+    return store.get(ref);
+  }
+
+  /// Returns an [ImpulseReference] representing this singleton.
+  @override
+  ImpulseReference<T> call() {
+    return ImpulseReference(
+      key: this,
+      create: create,
+      isFactory: false,
+      keepAlive: false,
+      reassemble: reassemble,
+      dispose: dispose,
+    );
+  }
+}
+
+/// A reference that creates and caches a single instance in the store forever (until the store gets reset).
+/// Commonly this is called a singleton.
+@immutable
+class SingletonRef<T> implements KeylessRef<T> {
+  /// Creates a [Ref] with a [create] function.
+  const SingletonRef(this.create, {this.reassemble, this.dispose});
+
+  /// The function used to create the object.
+  final Create<T> create;
+
+  /// The function used to update the object when the box reassembles.
+  ///
+  /// This usefull to ensure compatibility with hot reload.
+  final Update<T>? reassemble;
+
+  /// An optional function to handle manual disposal of the object.
+  final Dispose<T>? dispose;
 
   /// Get the value this [Ref] refers to from the [store].
   T get(Store store) {
@@ -122,12 +158,13 @@ class Ref<T> {
   }
 
   /// Returns an [ImpulseReference] representing this singleton.
+  @override
   ImpulseReference<T> call() {
     return ImpulseReference(
       key: this,
       create: create,
       isFactory: false,
-      keepAlive: keepAlive,
+      keepAlive: true,
       reassemble: reassemble,
       dispose: dispose,
     );
@@ -138,13 +175,7 @@ class Ref<T> {
 @immutable
 class FamilyRef<T, R> {
   /// Creates a [FamilyRef] with a [create] function.
-  /// If [keepAlive] is true, the objects will not be automatically dropped when not in use.
-  const FamilyRef(
-    this.create, {
-    this.reassemble,
-    this.dispose,
-    this.keepAlive = false,
-  });
+  const FamilyRef(this.create, {this.reassemble, this.dispose});
 
   /// The function used to create the object with a parameter.
   final FamilyCreate<T, R> create;
@@ -156,9 +187,6 @@ class FamilyRef<T, R> {
 
   /// An optional function to handle manual disposal of the object.
   final Dispose<T>? dispose;
-
-  /// Whether the objects should remain in the store even when their reference count reaches zero.
-  final bool keepAlive;
 
   /// Get the value this [FamilyRef] refers to from the [store].
   T get(Store store, R param) {
@@ -173,7 +201,7 @@ class FamilyRef<T, R> {
       key: (this, param),
       create: (store) => create(store, param),
       isFactory: false,
-      keepAlive: keepAlive,
+      keepAlive: false,
       reassemble: reassemble,
       dispose: dispose,
     );