+++ packages/impulse_flutter/lib/src/context_extensions.dart
@@ -8,8 +8,7 @@ import 'store_scope.dart';
 extension GetContext on BuildContext {
   /// Retrieves the value of a [ref] from the nearest [StoreScope].
   /// This does not create a dependency, so the widget will not rebuild when the value changes.
-  /// Should NOT be called during the build phase.
-  T peek<T>(ImpulseReference<T> ref) {
+  T read<T>(ImpulseReference<T> ref) {
     final box = StoreScope.box(this, ref, depend: false);
 
     return box.produce();
@@ -20,8 +19,7 @@ extension GetContext on BuildContext {
 extension BindContext on BuildContext {
   /// Binds the current widget to a [ref] from the nearest [StoreScope].
   /// The widget will automatically rebuild whenever the reference notifies of a change.
-  /// Must only be called during the build phase.
-  T dependOn<T>(ImpulseReference<T> ref) {
+  T bind<T>(ImpulseReference<T> ref) {
     final box = StoreScope.box(this, ref, depend: true);
 
     return box.produce();