+++ packages/impulse_flutter/lib/src/flutter_delegate.dart
@@ -2,7 +2,10 @@ import 'package:flutter/foundation.dart';
 import 'package:impulse/impulse.dart' hide $store;
 
 /// The global default [Store] instance configured for Flutter.
-final $store = Store(delegate: FlutterReactivityDelegate());
+final $store = createStore();
+
+/// Creates a [Store] configured for flutter.
+Store createStore() => Store(delegate: FlutterReactivityDelegate());
 
 /// A [ReactivityDelegate] pre-configured with [FlutterAdapter] to support [Listenable] and [ChangeNotifier].
 class FlutterReactivityDelegate<T> extends ReactivityDelegate {