+++ packages/impulse_signals/lib/impulse_signals.dart
@@ -1,7 +1,9 @@
 /// Signals integration for the Impulse state management library.
 library;
 
-export 'package:impulse_flutter/impulse_flutter.dart';
+export 'package:impulse_flutter/impulse_flutter.dart'
+    hide Result, MapResult, attempt;
+
 export 'package:signals_flutter/signals_flutter.dart';
 export 'src/controller.dart';
 export 'src/attempt.dart';