+++ packages/impulse/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- Added error handling method `attempt` and `Result<T>` record typedef.
+- Added `select` method to the store to react only to specific value changes.
+- Made `watch` and `select` use reference counting for consitency.
+- Made documentation improvements
+
 ## 0.1.1
 
 - Minor readme improvements